backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
python-dotenv==1.0.0
flask-jwt-extended
flask-cors==4.0.0
psycopg2-binary